requirements-test.txt
pytest==7.4.0
pytest-cov==4.1.0
python-dotenv==1.0.0
spacy==3.7.2
sentence-transformers==2.2.2
neo4j==5.14.1
qdrant-client==1.6.4
pymongo==4.6.1
nltk==3.8.1
scikit-learn==1.3.0
numpy>=1.24.0
tqdm>=4.65.0
huggingface-hub==0.20.3
transformers==4.37.2
torch>=2.1.0 